Hoist DefaultObjectTracer{,Factory}
Move it one level above the module hierarchy
