+++ src/util/reference_processor.rs
@@ -5,7 +5,7 @@ use std::sync::Mutex;
 use std::vec::Vec;
 
 use crate::plan::is_nursery_gc;
-use crate::plan::tracing::gc_work::weakref::DefaultObjectTracerContext;
+use crate::plan::tracing::gc_work::DefaultObjectTracerContext;
 use crate::plan::tracing::Trace;
 use crate::scheduler::WorkBucketStage;
 use crate::util::ObjectReference;
